REF: Drop React
