+++ app/index.html
@@ -1,10 +1,21 @@
 <!DOCTYPE html>
 <html>
-    <head>
-        <meta charset="utf-8">
-        <title>QIIME 2 Visualization Viewer</title>
-    </head>
-    <body>
-        <div id="root"></div>
-    </body>
+<head>
+    <meta charset="utf-8">
+    <title>QIIME 2 Visualization Viewer</title>
+</head>
+<body>
+    <div class="container-fluid">
+        <div class="row">
+            <img id="logo" />
+        </div>
+        <div class="row">
+            <div class="form-group">
+                <form action="#" method="post">
+                    <input id="fileinput" type="file" accept=".qzv"/>
+                </form>
+            </div>
+        </div>
+    </div>
+</body>
 </html>
