+++ package.json
@@ -7,8 +7,7 @@
     "lint:css": "stylelint 'app/**/*.css' --config config/stylelint.yaml",
     "lint:js": "eslint app/ config/ --ext .js,.jsx --config config/eslint.yaml",
     "lint": "npm run lint:js && npm run lint:css",
-    "pretest": "npm run lint --silent",
-    "test": "mocha --compilers js:babel-core/register --recursive",
+    "test": "npm run lint",
     "start": "webpack-dev-server --config config/webpack.dev.config.js --host localhost --port 4242 --inline --hot",
     "build": "webpack --config config/webpack.prod.config.js"
   },
@@ -27,8 +26,6 @@
     "extract-text-webpack-plugin": "^1.0.1",
     "html-webpack-plugin": "^2.22.0",
     "jszip": "^3.1.2",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
     "serviceworker-webpack-plugin": "^0.1.5",
     "webpack": "^1.13.2"
   },
@@ -38,17 +35,10 @@
     "babel-loader": "^6.2.5",
     "babel-plugin-transform-object-rest-spread": "^6.16.0",
     "babel-preset-es2015": "^6.16.0",
-    "babel-preset-react": "^6.16.0",
-    "chai": "^3.5.0",
     "css-loader": "^0.25.0",
     "eslint": "^3.7.0",
     "eslint-config-airbnb": "^12.0.0",
-    "eslint-plugin-import": "^1.16.0",
-    "eslint-plugin-jsx-a11y": "^2.2.2",
-    "eslint-plugin-react": "^6.3.0",
     "file-loader": "^0.9.0",
-    "mocha": "^3.1.0",
-    "postcss-loader": "^0.13.0",
     "style-loader": "^0.13.1",
     "stylelint": "^7.3.1",
     "stylelint-config-standard": "^13.0.2",
@@ -56,13 +46,8 @@
     "webpack-dev-server": "^1.16.1"
   },
   "babel": {
-    "presets": [
-      "es2015",
-      "react"
-    ],
-    "plugins": [
-      "transform-object-rest-spread"
-    ]
+    "presets": ["es2015"],
+    "plugins": ["transform-object-rest-spread"]
   },
   "private": true
 }
