--- test/app_spec.js
@@ -1,7 +1,0 @@
-import { expect } from 'chai';
-
-describe('tests', () => {
-    it('passes', () => {
-        expect(1).to.equal(1);
-    })
-});
